+++ Rules/IPv4/Inbound/Temporary.ps1
@@ -94,7 +94,7 @@ if ($Develop)
 		@Logs | Format-Output @Logs
 
 	$mDnsUsers = Get-SDDL -Domain "NT AUTHORITY" -User "NETWORK SERVICE" @Logs
-	Merge-SDDL ([ref] $mDnsUsers) (Get-SDDL -Group "Users") @Logs
+	Merge-SDDL ([ref] $mDnsUsers) $UsersGroupSDDL @Logs
 
 	# NOTE: should be network service
 	New-NetFirewallRule -DisplayName "Troubleshoot UDP - mDNS" `