+++ Rules/IPv4/Outbound/WindowsSystem.ps1
@@ -565,7 +565,7 @@ In order to target builds to your machine, we need to know a few important thing
 
 # TODO: USOAccounts testing with users, should be SYSTEM
 $USOAccounts = Get-SDDL -Domain "NT AUTHORITY" -User "SYSTEM" @Logs
-Merge-SDDL ([ref] $USOAccounts) (Get-SDDL -Group "Users") @Logs
+Merge-SDDL ([ref] $USOAccounts) $UsersGroupSDDL @Logs
 
 # TODO: Not available in Windows Server 2019
 $Program = "%SystemRoot%\System32\usocoreworker.exe"